Fix stuff in kimden#45 PR first round of code review
