+++ src/race/race_manager.hpp
@@ -34,7 +34,7 @@
 
 #include "network/remote_kart_info.hpp"
 #include "race/grand_prix_data.hpp"
-#include "race/itempolicy.hpp"
+#include "race/item_policy.hpp"
 #include "utils/vec3.hpp"
 #include "utils/types.hpp"
 
@@ -251,7 +251,7 @@ private:
 
 public:
     ItemPolicy m_item_policy;
-    void setItemPolicy(std::string str) { m_item_policy.fromString(str); m_item_policy.m_leader_section = -1; m_item_policy.m_virtualpace_code = -1; m_item_policy.m_restart_count = 0; } 
+    void setItemPolicy(std::string str); 
     ItemPolicy *getItemPolicy() { return &m_item_policy; };
 
     /** This data structure accumulates kart data and race result data from
