+++ src/utils/lobby_settings.cpp
@@ -78,7 +78,7 @@ void LobbySettings::setupContextUser()
     loadWhiteList();
     loadPreservedSettings();
 
-    RaceManager::get()->setItemPolicy(ServerConfig::m_itempolicy);
+    RaceManager::get()->setItemPolicy(ServerConfig::m_item_policy);
 
 
     m_live_players = ServerConfig::m_live_players;
